commit-sayang<3

+++ database/migrations/2024_09_15_104507_create_riwayat_pembayarans_table.php
@@ -16,7 +16,7 @@ class CreateRiwayatPembayaransTable extends Migration
             $table->unsignedBigInteger('id_vehicles');
             $table->string('kode_kendaraan');
             $table->date('tanggal_bayar');
-            $table->decimal('total_bayar', 10, 2);
+            $table->integer('total_bayar');
             $table->string('bukti_pembayaran');
             $table->boolean('konfirmasi_pembayaran')->default(false);
             $table->timestamps();
